Adjust the SSJ navigation structure
